COMMERCE-5450 utilities folder renamed to utils
