+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/resources/META-INF/resources/data_set_display/data_renderers/DefaultRenderer.js
@@ -16,7 +16,7 @@ import ClayIcon from '@clayui/icon';
 import PropTypes from 'prop-types';
 import React from 'react';
 
-import {logError} from '../utilities/logError';
+import {logError} from '../utils/logError';
 import TooltipTextRenderer from './TooltipTextRenderer';
 
 function DefaultRenderer({value}) {
