+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/resources/META-INF/resources/data_set_display/data_renderers/ImageRenderer.js
@@ -16,7 +16,7 @@ import ClaySticker from '@clayui/sticker';
 import PropType from 'prop-types';
 import React from 'react';
 
-import {getValueFromItem} from '../utilities/index';
+import {getValueFromItem} from '../utils/index';
 
 function ImageRenderer(props) {
 	const imageProps =
