+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/resources/META-INF/resources/data_set_display/management_bar/components/filters/DateRangeFilter.js
@@ -23,7 +23,7 @@ import {
 	formatDateObject,
 	formatDateRangeObject,
 	getDateFromDateString,
-} from '../../../utilities/dates';
+} from '../../../utils/dates';
 
 function getOdataString(value, key) {
 	if (value.from && value.to) {
