+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/resources/META-INF/resources/data_set_display/modal/Modal.js
@@ -21,11 +21,11 @@ import {
 	CLOSE_MODAL,
 	IS_LOADING_MODAL,
 	OPEN_MODAL,
-} from '../utilities/eventsDefinitions';
-import {isPageInIframe} from '../utilities/iframes';
-import {liferayNavigate} from '../utilities/index';
-import {INITIAL_MODAL_SIZE} from '../utilities/modals/constants';
-import {resolveModalHeight} from '../utilities/modals/index';
+} from '../utils/eventsDefinitions';
+import {isPageInIframe} from '../utils/iframes';
+import {liferayNavigate} from '../utils/index';
+import {INITIAL_MODAL_SIZE} from '../utils/modals/constants';
+import {resolveModalHeight} from '../utils/modals/index';
 
 function Modal({
 	id,
