+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/resources/META-INF/resources/data_set_display/thunks/persistActiveView.js
@@ -13,7 +13,7 @@
  */
 
 import {updateActiveView} from '../actions/updateActiveView';
-import {saveViewSettings} from '../utilities/saveViewSettings';
+import {saveViewSettings} from '../utils/saveViewSettings';
 
 export default function persistActiveView({
 	activeViewName,
