+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/resources/META-INF/resources/data_set_display/thunks/persistVisibleFieldNames.js
@@ -13,7 +13,7 @@
  */
 
 import {updateVisibleFieldNames} from '../actions/updateVisibleFieldNames';
-import {saveViewSettings} from '../utilities/saveViewSettings';
+import {saveViewSettings} from '../utils/saveViewSettings';
 
 export default function persistVisibleFieldNames({
 	appURL,
