+++ modules/apps/frontend-taglib/frontend-taglib-clay/src/main/resources/META-INF/resources/data_set_display/views/table/TableCell.js
@@ -21,7 +21,7 @@ import {
 	getDataRendererById,
 	getDataRendererByURL,
 	getInputRendererById,
-} from '../../utilities/dataRenderers';
+} from '../../utils/dataRenderers';
 
 function InlineEditInputRenderer({
 	itemId,
